pyproject.toml
[tool.poetry]
name = "magoji"
version = "0.1.0"
description = "A shared utility Discord bot"
authors = ["vcokltfre <vcokltfre@gmail.com>", "chili <chili@github.com>", "anand <anand2312@github.com>", "ToxicKidz <ToxicKidz@github.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "3.9"
"discord.py" = "1.7.1"
pyyaml = "*"
python-dotenv = "*"
aiohttp = "*"
aiodns = "*"
asyncpg = "*"
jishaku = "*"
discord-ext-menus = {git = "https://github.com/Rapptz/discord-ext-menus.git"}
GitPython = "^3.1.13"
psutil = "^5.8.0"

[tool.poetry.dev-dependencies]
pre-commit = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
